modification creation dépense

+++ back/src/controllers/depense.controller.ts
@@ -101,15 +101,17 @@ export const creer_depense: RequestHandler = async (req: Request, res: Response)
     // On récupère cette information depuis le middleware JWT
     const utilisateur_id: number = (<any>req).user;
 
-    const { groupe_id_param, titre, montant, participants } = req.body;
+    const { groupe_id_param, titre, montant, url_image, participant_payeur_id, participants } = req.body;
 
-    if (!["number"].includes(typeof groupe_id_param) || !["string"].includes(typeof titre) || titre == "" || !["number"].includes(typeof montant) || !Array.isArray(participants))
+    if (!["number"].includes(typeof groupe_id_param) || !["string"].includes(typeof titre) || titre == "" || (url_image != null && !["string"].includes(typeof url_image)) || !["number"].includes(typeof participant_payeur_id) || !["number"].includes(typeof montant) || !Array.isArray(participants))
         return res.status(http_response_util.statuts.erreur_client.parametres_manquant).json({
             message: "Un ou plusieurs paramètres ne sont pas présent dans la requête.",
             parametres: [
                 { nom: "groupe_id_param", type: "string", facultatif: false },
                 { nom: "titre", type: "string", facultatif: false },
                 { nom: "montant", type: "float", facultatif: false },
+                { nom: "url_image", type: "string", facultatif: true },
+                { nom: "participant_payeur_id", type: "number", facultatif: false },
                 { nom: "participants", type: "array", facultatif: false }
             ]
         });
@@ -122,6 +124,16 @@ export const creer_depense: RequestHandler = async (req: Request, res: Response)
     if (isNaN(Number(montant)))
         return res.status(http_response_util.statuts.erreur_client.contenu_pas_autorise).json({
             message: "Le montant doit être un nombre."
+        });
+
+    if (Number(montant) <= 0)
+        return res.status(http_response_util.statuts.erreur_client.contenu_pas_autorise).json({
+            message: "Le montant doit être supérieur à 0."
+        });
+
+    if (isNaN(Number(participant_payeur_id)))
+        return res.status(http_response_util.statuts.erreur_client.contenu_pas_autorise).json({
+            message: "Le paye_par doit être un nombre."
         });
 
     const groupe_id: number = Number(groupe_id_param);
@@ -159,33 +171,51 @@ export const creer_depense: RequestHandler = async (req: Request, res: Response)
         });
     }
 
-    let participants_existant: Participant_Groupe | null = null;
+    let participants_groupe: Participant_Groupe[] = [];
 
     try {
-        const participants: Participant_Groupe[] = await groupe_service.recuperer_participants(groupe_id);
-
-        participants_existant = participants.find((participant: Participant_Groupe) => participant.fk_utilisateur_id == utilisateur_id);
-
-        if (!participants_existant)
-            return res.status(http_response_util.statuts.erreur_client.mauvaise_requete).json({
-                message: "Vous n'êtes pas affilié à ce groupe."
-            });
-
-        if (participants_existant.quitte_le != null)
-            return res.status(http_response_util.statuts.erreur_client.mauvaise_requete).json({
-                message: "Vous avez quitté ce groupe."
-            });
-    } catch (error: PrismaClientKnownRequestError | any) {
+        participants_groupe = await groupe_service.recuperer_participants(groupe_id);
+    } catch (error) {
         return res.status(http_response_util.statuts.erreur_serveur.erreur_interne).json({
-            message: "Une erreur serveur est survenue lors de la récupération des participants du groupe partagé.",
-            erreur: error
+            message: "Une erreur serveur est survenue lors de la récupération de tous les participants du groupe."
         });
     }
+
+    let participant_existant: Participant_Groupe | null = participants_groupe.find((participant: Participant_Groupe) => participant.fk_utilisateur_id == utilisateur_id);
+
+    if (!participant_existant)
+        return res.status(http_response_util.statuts.erreur_client.mauvaise_requete).json({
+            message: "Vous n'êtes pas affilié à ce groupe."
+        });
+
+    if (participant_existant.quitte_le != null)
+        return res.status(http_response_util.statuts.erreur_client.mauvaise_requete).json({
+            message: "Vous avez quitté ce groupe."
+        });
+
+    if (!participant_existant.peut_creer_depense)
+        return res.status(http_response_util.statuts.erreur_client.mauvaise_requete).json({
+            message: "Vous n'avez pas les permissions pour créer une dépense dans ce groupe."
+        });
+
+    let participant_payeur: Participant_Groupe | null = participants_groupe.find((participant: Participant_Groupe) => participant.pk_participant_groupe_id == Number(participant_payeur_id));
+
+    if (!participant_payeur)
+        return res.status(http_response_util.statuts.erreur_client.mauvaise_requete).json({
+            message: "Le participant payeur n'existe pas."
+        });
+
+    participants.map((p: { fk_participant_groupe_id: number, montant: number | null }) => {
+        if(!participants_groupe.find((participant_groupe: Participant_Groupe) => participant_groupe.pk_participant_groupe_id == p.fk_participant_groupe_id))
+            return res.status(http_response_util.statuts.erreur_client.mauvaise_requete).json({
+                message: "Le participant n°" + p.fk_participant_groupe_id + " n'existe pas."
+            });
+    });
 
     let depense: Depense | null = null;
 
     try {
-        depense = await depense_service.creer_depense(groupe_id, participants_existant.pk_participant_groupe_id, titre, montant);
+        depense = await depense_service.creer_depense(groupe_existant.pk_groupe_id, participant_payeur.pk_participant_groupe_id, titre, montant, url_image);
     } catch (error: PrismaClientKnownRequestError | any) {
         return res.status(http_response_util.statuts.erreur_serveur.erreur_interne).json({
             message: "Une erreur serveur est survenue lors de la création de la dépense.",
@@ -204,19 +234,12 @@ export const creer_depense: RequestHandler = async (req: Request, res: Response)
         });
     }
 
-    if (!participants.find((participant: { fk_participant_groupe_id: number, montant: number | null }) => participant.fk_participant_groupe_id == participants_existant.pk_participant_groupe_id)) {
-        try {
-            await depense_service.lier_depense_participants(depense.pk_depense_id, participants_existant.pk_participant_groupe_id, null);
-        } catch (error: PrismaClientKnownRequestError | any) {
-            return res.status(http_response_util.statuts.erreur_serveur.erreur_interne).json({
-                message: "Une erreur serveur est survenue lors de la liaison de la dépense avec les participants.",
-                erreur: error
-            });
-        }
-    }
     return res.status(http_response_util.statuts.succes.ok).json({
         message: "La dépense a été créée avec succès.",
-        data: depense
+        data: {
+            ...depense,
+            ajoute_le: moment(depense.ajoute_le).format(moment_date_time_format)
+        }
     });
 }
 
@@ -584,10 +607,14 @@ export const supprimer_depense: RequestHandler = async (req: Request, res: Respo
         });
     }
 
+    let participants_existant: Participant_Groupe | null = null;
+
     try {
         const participants: Participant_Groupe[] = await groupe_service.recuperer_participants(depense_existant.fk_groupe_id);
 
-        if (!participants.find((participant: Participant_Groupe) => participant.fk_utilisateur_id == utilisateur_id && participant.quitte_le == null))
+        participants_existant = participants.find((participant: Participant_Groupe) => participant.fk_utilisateur_id == utilisateur_id && participant.quitte_le == null);
+
+        if (!participants_existant)
             return res.status(http_response_util.statuts.erreur_client.mauvaise_requete).json({
                 message: "Vous n'êtes pas affilié à ce groupe."
             });
@@ -597,6 +624,11 @@ export const supprimer_depense: RequestHandler = async (req: Request, res: Respo
             erreur: error
         });
     }
+
+    if (!participants_existant.peut_supprimer_depense)
+        return res.status(http_response_util.statuts.erreur_client.mauvaise_requete).json({
+            message: "Vous n'avez pas les permissions pour supprimer une dépense dans ce groupe."
+        });
 
     try {
         await depense_service.supprimer_depense(depense_id);
@@ -610,4 +642,150 @@ export const supprimer_depense: RequestHandler = async (req: Request, res: Respo
             erreur: error
         });
     }
+}
+
+export const modifier_depense: RequestHandler = async (req: Request, res: Response) => {
+    // On récupère cette information depuis le middleware JWT
+    const utilisateur_id: number = (<any>req).user;
+
+    const { depense_id, titre, montant, url_image } = req.body;
+
+    if (!["number"].includes(typeof depense_id) || !["string"].includes(typeof titre) || titre == "" || !["number"].includes(typeof montant) || (url_image && !["string"].includes(typeof url_image)))
+        return res.status(http_response_util.statuts.erreur_client.parametres_manquant).json({
+            message: "Un ou plusieurs paramètres ne sont pas présent dans la requête.",
+            parametres: [
+                { nom: "depense_id", type: "string", facultatif: false },
+                { nom: "titre", type: "string", facultatif: false },
+                { nom: "montant", type: "float", facultatif: false },
+                { nom: "url_image", type: "string", facultatif: true }
+            ]
+        });
+
+    if (isNaN(Number(depense_id)))
+        return res.status(http_response_util.statuts.erreur_client.contenu_pas_autorise).json({
+            message: "La depense_id doit être un nombre."
+        });
+
+    if (isNaN(Number(montant)))
+        return res.status(http_response_util.statuts.erreur_client.contenu_pas_autorise).json({
+            message: "Le montant doit être un nombre."
+        });
+
+    if (Number(montant) <= 0)
+        return res.status(http_response_util.statuts.erreur_client.contenu_pas_autorise).json({
+            message: "Le montant doit être supérieur à 0."
+        });
+
+    let utilisateur_existant: Utilisateur | null = null;
+
+    // On cherche l'utilisateur par son id avec son token JWT
+    try {
+        utilisateur_existant = await utilisateur_service.recuperer_utilisateur_par_id(utilisateur_id);
+
+        if (!utilisateur_existant)
+            return res.status(http_response_util.statuts.erreur_client.mauvaise_requete).json({
+                message: "Malgré votre token jwt, nous ne vous avons pas trouvé dans la base de données."
+            });
+    } catch (error: PrismaClientKnownRequestError | any) {
+        return res.status(http_response_util.statuts.erreur_serveur.erreur_interne).json({
+            message: "Une erreur serveur est survenue lors de la récupération des données de l'utilisateur.",
+            erreur: error
+        });
+    }
+
+    let depense_existante: Depense | null = null;
+
+    try {
+        depense_existante = await depense_service.recuperer_depense_par_id(Number(depense_id));
+
+        if (!depense_existante)
+            return res.status(http_response_util.statuts.erreur_client.mauvaise_requete).json({
+                message: "La dépense n'existe pas."
+            });
+    } catch (error: PrismaClientKnownRequestError | any) {
+        return res.status(http_response_util.statuts.erreur_serveur.erreur_interne).json({ 
+            message: "Une erreur serveur est survenue lors de la récupération de la dépense.",
+            erreur: error
+        });
+    }
+    
+
+    let groupe_existant: Groupe | null = null;
+
+    try {
+        groupe_existant = await groupe_service.recuperer_groupe(depense_existante.fk_groupe_id);
+
+        if (!groupe_existant)
+            return res.status(http_response_util.statuts.erreur_client.mauvaise_requete).json({
+                message: "Le groupe partagé n'existe pas."
+            });
+    } catch (error: PrismaClientKnownRequestError | any) {
+        return res.status(http_response_util.statuts.erreur_serveur.erreur_interne).json({
+            message: "Une erreur serveur est survenue lors de la récupération des informations du groupe.",
+            erreur: error
+        });
+    }
+
+    let participants_existant: Participant_Groupe | null = null;
+
+    try {
+        const participants: Participant_Groupe[] = await groupe_service.recuperer_participants(groupe_existant.pk_groupe_id);
+
+        participants_existant = participants.find((participant: Participant_Groupe) => participant.fk_utilisateur_id == utilisateur_id);
+
+        if (!participants_existant)
+            return res.status(http_response_util.statuts.erreur_client.mauvaise_requete).json({
+                message: "Vous n'êtes pas affilié à ce groupe."
+            });
+
+        if (participants_existant.quitte_le != null)
+            return res.status(http_response_util.statuts.erreur_client.mauvaise_requete).json({
+                message: "Vous avez quitté ce groupe."
+            });
+    } catch (error: PrismaClientKnownRequestError | any) {
+        return res.status(http_response_util.statuts.erreur_serveur.erreur_interne).json({
+            message: "Une erreur serveur est survenue lors de la récupération des participants du groupe partagé.",
+            erreur: error
+        });
+    }
+
+    if (!participants_existant.peut_modifier_depense)
+        return res.status(http_response_util.statuts.erreur_client.mauvaise_requete).json({
+            message: "Vous n'avez pas les permissions pour modifier une dépense dans ce groupe."
+        });
+
+    let depenses_groupe: Depense[] = [];
+
+    try {
+        depenses_groupe = await depense_service.recuperer_toutes_depenses(groupe_existant.pk_groupe_id);
+    } catch (error) {
+        return res.status(http_response_util.statuts.erreur_serveur.erreur_interne).json({ 
+            message: "Une erreur serveur est survenue lors de la récupération des dépenses du groupe.",
+            erreur: error
+        });
+    }
+
+    if (depenses_groupe.filter((d: Depense) => d.montant != null).reduce((somme: number, d: Depense) => somme + d.montant, 0) > Number(montant))
+        return res.status(http_response_util.statuts.erreur_client.mauvaise_requete).json({
+            message: "Le montant de la dépense ne peut pas être inférieur à la somme des montants des autres dépenses du groupe."
+        });
+
+    let depense_modifiee: Depense | null = null;
+
+    try {
+        depense_modifiee = await depense_service.modifier_depense(depense_id, titre, montant, url_image);
+    } catch (error: PrismaClientKnownRequestError | any) {
+        return res.status(http_response_util.statuts.erreur_serveur.erreur_interne).json({ 
+            message: "Une erreur serveur est survenue lors de la modification de la dépense.",
+            erreur: error
+        });
+    }
+
+    return res.status(http_response_util.statuts.succes.ok).json({
+        message: "La dépense a été modifiée.",
+        data: {
+            ...depense_modifiee,
+            ajoute_le: moment(depense_modifiee.ajoute_le).format(moment_date_time_format)
+        }
+    });
 }
